refactor!: change design of SDK completly
