--- src/main/java/org/chocodev/Error/SDK/FieldException.java
@@ -1,7 +1,0 @@
-package org.chocodev.Error.SDK;
-
-public class FieldException extends SDKException {
-    public FieldException(String msg) {
-        super(msg);
-    }
-}
