--- src/main/java/org/chocodev/Error/SDK/MapperException.java
@@ -1,7 +1,0 @@
-package org.chocodev.Error.SDK;
-
-public class MapperException extends SDKException {
-    public MapperException(String msg){
-        super(msg);
-    }
-}
