--- src/main/java/org/chocodev/Error/SDK/SDKException.java
@@ -1,7 +1,0 @@
-package org.chocodev.Error.SDK;
-
-public class SDKException extends RuntimeException {
-    public SDKException(String msg){
-        super(msg);
-    }
-}
