--- src/main/java/org/chocodev/Error/UTApi/RequestException.java
@@ -1,26 +1,0 @@
-package org.chocodev.Error.UTApi;
-
-public class RequestException extends UTApiException {
-    private int status;
-
-    public int getStatusCode() {
-        return this.status;
-    }
-
-    public RequestException(String msg) {
-        super(msg);
-    }
-
-    public RequestException(RequestException exception) {
-        this(exception.getMessage(), exception.getStatusCode());
-    }
-
-    public RequestException(Exception exception) {
-        this(exception.getMessage());
-    }
-
-    public RequestException(String body, int statusCode) {
-        this(body);
-        this.status = statusCode;
-    }
-}
