--- src/main/java/org/chocodev/Error/UTApi/UTApiException.java
@@ -1,7 +1,0 @@
-package org.chocodev.Error.UTApi;
-
-public class UTApiException extends Exception {
-    public UTApiException(String msg) {
-        super(msg);
-    }
-}
