--- src/main/java/org/chocodev/Fetch/IFetch.java
@@ -1,7 +1,0 @@
-package org.chocodev.Fetch;
-
-import org.chocodev.Error.UTApi.RequestException;
-
-public interface IFetch<Response> {
-    public Response request() throws RequestException;
-}
