--- src/main/java/org/chocodev/Requests/RequestMapper.java
@@ -1,11 +1,0 @@
-package org.chocodev.Requests;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public class RequestMapper {
-    public static ObjectMapper objectMapper = new ObjectMapper();
-
-    private RequestMapper() {
-
-    }
-}
