--- src/main/java/org/chocodev/UTApiConfig.java
@@ -1,9 +1,0 @@
-package org.chocodev;
-
-
-public class UTApiConfig {
-    private UTApiConfig(){};
-    public final static String UTFileName = "filename";
-    public final static String accessFileUrl = "https://utfs.io";
-    public final static String deleteFileUrl = "https://api.uploadthing.com/v6/deleteFiles";
-}
