--- src/main/java/org/chocodev/UploadThing/Constants/ContentDisposition.java
@@ -1,5 +1,0 @@
-package org.chocodev.UploadThing.Constants;
-
-public enum ContentDisposition {
-    
-}
