--- src/main/java/org/chocodev/UploadThing/Constants/Messages.java
@@ -1,13 +1,0 @@
-package org.chocodev.UploadThing.Constants;
-
-public class Messages {
-    private Messages() {
-
-    }
-    public static final String mapperErrorMessage = "";
-    public static final String customizerErrorMessage = "";
-    public static final String fieldErrorMessage = "";
-    public static String requestErrorMessage(String msg){
-        return msg + "";
-    }
-}
