--- src/main/java/org/chocodev/UploadThing/Customizer/ICustomizer.java
@@ -1,7 +1,0 @@
-package org.chocodev.UploadThing.Customizer;
-
-public interface ICustomizer<TObject> {
-    public ICustomizer<TObject> builder();
-    public TObject build();
-    public TObject withDefault();
-}
