--- src/main/java/org/chocodev/UploadThing/File/IFile.java
@@ -1,8 +1,0 @@
-package org.chocodev.UploadThing.File;
-
-public interface IFile {
-  public byte[] getBytes();
-  public String getName();
-  public String getType();
-  public String getFileExtension();
-}
