--- src/main/java/org/chocodev/UploadThing/File/IUploadThingFile.java
@@ -1,7 +1,0 @@
-package org.chocodev.UploadThing.File;
-
-public interface IUploadThingFile {
-    public String getKey();
-    public String getFileName();
-    public String getCustomId();
-}
