--- src/main/java/org/chocodev/UploadThing/Status/IStatus.java
@@ -1,5 +1,0 @@
-package org.chocodev.UploadThing.Status;
-
-public interface IStatus {
-    public boolean isSuccess();
-}
