+++ src/main/java/org/chocodev/core/File/IUTFile.java
@@ -1,0 +1,8 @@
+package org.chocodev.core.File;
+
+public interface IUTFile {
+    public String getCustomId();
+    public String getFileKey();
+    public String getFileName();
+
+}
