+++ src/main/java/org/chocodev/internal/Exceptions/SDK/BadApiCallException.java
@@ -1,0 +1,9 @@
+package org.chocodev.internal.Exceptions.SDK;
+
+public class BadApiCallException extends SDKException {
+
+    public BadApiCallException(String msg) {
+        super(msg);
+    }
+    
+}
