+++ src/main/java/org/chocodev/internal/Exceptions/SDK/CustomizerException.java
@@ -1,0 +1,13 @@
+package org.chocodev.internal.Exceptions.SDK;
+
+import org.chocodev.util.Constants.Messages;
+
+public class CustomizerException extends SDKException {
+    public CustomizerException(){
+        super(Messages.customizerErrorMessage);
+    }
+    public CustomizerException(String msg) {
+        super(msg);
+    }
+    
+}
