+++ src/main/java/org/chocodev/internal/Exceptions/SDK/FieldException.java
@@ -1,0 +1,7 @@
+package org.chocodev.internal.Exceptions.SDK;
+
+public class FieldException extends SDKException {
+    public FieldException(String msg) {
+        super(msg);
+    }
+}
