+++ src/main/java/org/chocodev/internal/Exceptions/SDK/MapperException.java
@@ -1,0 +1,7 @@
+package org.chocodev.internal.Exceptions.SDK;
+
+public class MapperException extends SDKException {
+    public MapperException(String msg){
+        super(msg);
+    }
+}
