+++ src/main/java/org/chocodev/internal/Exceptions/SDK/SDKException.java
@@ -1,0 +1,7 @@
+package org.chocodev.internal.Exceptions.SDK;
+
+public class SDKException extends RuntimeException {
+    public SDKException(String msg){
+        super(msg);
+    }
+}
