+++ src/main/java/org/chocodev/internal/Exceptions/UTApi/UTApiException.java
@@ -1,0 +1,7 @@
+package org.chocodev.internal.Exceptions.UTApi;
+
+public class UTApiException extends Exception {
+    public UTApiException(String msg) {
+        super(msg);
+    }
+}
