+++ src/main/java/org/chocodev/internal/UTApiConfig.java
@@ -1,0 +1,9 @@
+package org.chocodev.internal;
+
+
+public class UTApiConfig {
+    private UTApiConfig(){};
+    public final static String UTFileName = "filename";
+    public final static String accessFileUrl = "https://utfs.io";
+    public final static String deleteFileUrl = "https://api.uploadthing.com/v6/deleteFiles";
+}
