+++ src/main/java/org/chocodev/util/Constants/ContentDisposition.java
@@ -1,0 +1,5 @@
+package org.chocodev.util.Constants;
+
+public enum ContentDisposition {
+    
+}
