+++ src/main/java/org/chocodev/util/Constants/KeyType.java
@@ -1,0 +1,14 @@
+package org.chocodev.util.Constants;
+public enum KeyType {
+    FILE_KEY("fileKeys"),
+    CUSTOM_ID("customIds");
+
+    private final String value;
+    KeyType(String value) {
+        this.value = value;
+    }
+    public String getValue() {
+        return value;
+    }
+}
+
