+++ src/main/java/org/chocodev/util/Constants/Messages.java
@@ -1,0 +1,14 @@
+package org.chocodev.util.Constants;
+
+public class Messages {
+    private Messages() {
+
+    }
+    public static final String mapperErrorMessage = "";
+    public static final String customizerErrorMessage = "";
+    public static final String fieldErrorMessage = "";
+    public static final String noExceptionErrorMessage = "";
+    public static String requestErrorMessage(String msg){
+        return msg + "";
+    }
+}
